impletment my_plants view and user authentication rqmts

+++ app/plants/views.py
@@ -8,9 +8,10 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User                        # User signup
 from django.contrib.auth        import authenticate, login, logout # User login/logout
 
-from .models import Garden, Plant, Comment
+from .models import Garden, MyPlant, Plant, Comment
 from .forms  import UserSignupForm, UserLoginForm
-from .forms  import GardenAddUpdateForm, PlantAddUpdateForm, PlantCommentForm
+from .forms  import GardenAddUpdateForm, MyPlantAddForm
+from .forms  import PlantAddUpdateForm, PlantCommentForm
 
 import math
 import os
@@ -33,9 +34,10 @@ def index(request):
     """ Render the landing page for Gateway Gardens app """
     return render(request, 'plants/index.html')
 
-@login_required()
 def gardens_summary(request):
     """ Render the individual user's garden summary for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     gardens = Garden.objects.all()
     template = loader.get_template("plants/gardens_summary.html")
     context = {}
@@ -54,9 +56,10 @@ def gardens_summary(request):
         return HttpResponseRedirect(reverse('plants:gardens_add')) 
     return HttpResponse(template.render(context, request))
 
-@login_required()
 def gardens_add(request):
     """ Add garden to the database """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     garden = Garden()
     if request.POST:
         form = GardenAddUpdateForm(request.POST, request.FILES)
@@ -100,9 +103,10 @@ def gardens_add(request):
         context = { 'form' : form }
         return render(request, 'plants/gardens_add.html', context)
 
-@login_required()
 def gardens_update(request, id):
     """ Update garden details """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     garden = Garden.objects.get(id=id)
     if request.POST:
         form = GardenAddUpdateForm(request.POST, request.FILES)
@@ -186,20 +190,40 @@ def gardens_update(request, id):
                   }
         return render(request, 'plants/gardens_update.html', context)
  
-@login_required()
 def myplants_summary(request):
-    """ Render the My Plants page for Gateway Gardens app """
-    # AR: Under construction
-    plants = Plant.objects.all().values() 
-    # if Plant.objects.filter(gardens__garden_owner=request.user.username):
-    #     plants = Plant.objects.filter(gardens__garden_owner=request.user.username)
-    template = loader.get_template("plants/myplants_summary.html")
-    context = { "plants" : plants }
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
+    my_plants = MyPlant.objects.filter(owner = request.user.username)
+    template = loader.get_template('plants/myplants_summary.html')
+    context = { 'my_plants' : my_plants }
     return HttpResponse(template.render(context, request))
 
-@login_required()
+def myplants_add(request, id):
+    """ Associate a 'plant' to 'my_plant' """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
+    plant = Plant.objects.get(id=id)
+    my_plant = MyPlant()
+    if request.POST:
+        form = MyPlantAddForm(request.POST)
+        if form.is_valid():
+            my_plant.owner        = request.user.username                 #
+            my_plant.location     = form.cleaned_data.get("location")     #
+            my_plant.sun_exposure = form.cleaned_data.get("sun_exposure") #
+            my_plant.pH           = form.cleaned_data.get("pH")           #
+            my_plant.soil_type    = form.cleaned_data.get("soil_type")    #
+            my_plant.plant        = plant                                 # link my_plant to the specific plant
+            my_plant.save()
+        return HttpResponseRedirect(reverse('plants:index')) 
+    else:
+        form = MyPlantAddForm()
+        context = { 'form'  : form }
+        return render(request, 'plants/myplants_add.html', context)
+
 def plants_search(request):
     """ Render the Searchable summary list of plants with comments for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     if request.method == 'POST':                             # Executed when search is initiated
         template = loader.get_template("plants/plants_search.html")
         # Get the search criteria
@@ -313,9 +337,10 @@ def plants_search(request):
               }
     return HttpResponse(template.render(context, request))
 
-@login_required()
 def plants_details(request, id):
     """ Show a detailed view of a specific plant """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant.objects.get(id=id)                     # Uses the id to locate the correct record in the Plant table
     comments = Comment.objects.filter(plant__pk=id)      # get all comments related to the plant
     # format multiselect attributes to remove [, ', and ]
@@ -339,9 +364,10 @@ def plants_details(request, id):
             }
     return HttpResponse(template.render(context, request)) # Send "context" to template and output the html from the template
 
-@login_required()
 def plants_add(request):
     """ Render the page to add plants to the database for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant()
     if request.POST:
         form = PlantAddUpdateForm(request.POST, request.FILES)
@@ -396,9 +422,10 @@ def plants_add(request):
         context = { 'form' : form }
         return render(request, 'plants/plants_add.html', context)
 
-@login_required()
 def plants_update(request, id):
     """ Update the attributes for an existing plant """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant.objects.get(id=id)
     if request.POST:
         form = PlantAddUpdateForm(request.POST, request.FILES)
@@ -499,9 +526,10 @@ def plants_update(request, id):
         context = { 'form' : form }
         return render(request, 'plants/plants_update.html', context)
 
-@login_required()
 def plants_comment(request, id):
-    """ Add a comment to a plant """
+    """ Associate a comment to a plant """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant.objects.get(id=id)
     comment = Comment()
     if request.POST:
@@ -522,6 +550,8 @@ def plants_comment(request, id):
 
 def plant2garden(request, id):
     """ Add a plant to a garden """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant.objects.get(id=id)
     gardens = Garden.objects.all()
     for garden in gardens:
@@ -530,9 +560,10 @@ def plant2garden(request, id):
             plant.gardens.add(garden)                    # Associate the plant to the selected garden
     return HttpResponseRedirect(reverse('plants:plants_search'))
 
-@login_required()
 def plants_delete(request, id):
     """ Delete selected plant from the database  """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     plant = Plant.objects.get(id=id)
     if request.POST:
       plant.delete()
@@ -540,19 +571,22 @@ def plants_delete(request, id):
     context = {'plant': plant}
     return render(request, 'plants/plants_delete.html', context)
 
-@login_required
 def plants_glossary(request):
     """ Render the Glossary Page for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     return render(request, 'plants/plants_glossary.html')
 
-@login_required
 def plants_reference(request):
     """ Render the Reference Page for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     return render(request, 'plants/plants_reference.html')
 
-@login_required
 def plants_about(request):
     """ Render the About Page for Gateway Gardens app """
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('plants:user_login'))
     return render(request, 'plants/plants_about.html')
 
 def user_signup(request):
